support BIGINT, DATE and TIME data types

+++ src/parse_d.h
@@ -1,4 +1,6 @@
 /*
+ * parse_d.h - declarations of filed parsing functions
+ *
  * Copyright 2016 Guo, Xingchun <guoxingchun@gmail.com>
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -23,7 +25,7 @@
 #define VARCHAR_CUR_LEN_IND_BYTES 2
 #define NULL_VAL_IND_BYTES 2
 
-long parse_ixf_integer(const unsigned char *src, size_t bytes);
+long long parse_ixf_integer(const unsigned char *src, size_t bytes);
 _Bool column_is_null(const unsigned char *null_ind);
 char *decode_packed_decimal(char *buff, const unsigned char *src,
 			    size_t data_length);
